requirements.txt
streamlit
streamlit-lottie
pandas
matplotlib
altair
scipy
pytesseract
pillow

